requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
discord.py==2.6.3
khl.py==0.3.17
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
tencentcloud-sdk-python==3.0.1035
typing_extensions==4.15.0
yarl==1.20.1
# 新增翻译平台依赖
googletrans==4.0.0-rc1
baidu-aip==4.16.10
youdao-python==0.1.3
cryptography==41.0.3
requests==2.31.0